+++ spark-testkit/src/main/scala/geotrellis/spark/testkit/io/cog/COGPersistenceSpec.scala
@@ -19,6 +19,7 @@ package geotrellis.spark.testkit.io.cog
 import geotrellis.raster.CellGrid
 import geotrellis.raster.crop.TileCropMethods
 import geotrellis.raster.merge.TileMergeMethods
+import geotrellis.raster.resample._
 import geotrellis.raster.io.geotiff.compression.NoCompression
 import geotrellis.raster.io.geotiff._
 import geotrellis.raster.io.geotiff.reader.GeoTiffReader
@@ -145,6 +146,10 @@ abstract class COGPersistenceSpec[
         val tileSize = 1024
 
         writer.write[K, V](s"${layerId.name}-2", sample, layerId.zoom, keyIndexMethod, tileSize)
+      }
+
+      it("should create a COGLayer with the resample method set") {
+        COGLayer.fromLayerRDD(sample, layerId.zoom, NoCompression, options = Sum)
       }
 
       /*it("should delete a layer") {