+++ docs/CHANGELOG.rst
@@ -25,6 +25,7 @@ API Changes
   - **Change:** Replace `geotrellis.util.Functor` with `cats.Functor`
   - **Change:** Specifying the ``maxTileSize`` for a COGLayer that's to be written is now done via ``COGLayerWriter.Options``
     which can be passed directly to the ``write`` methods.
+  - **New:** The ``resampleMethod`` parameter has been added to ``COGLayerWriter.Options``.
 
 Fixes
 ^^^^^